Layout de resultado
Adicionado layout de resultado para matrizes

+++ app/src/main/java/com/example/matrizes/Celula.kt
@@ -1,4 +1,4 @@
 package com.example.matrizes
 
-data class Celula(var celula: Int?)
+data class Celula(var celula: Double?)
